requirements.txt
cryptography
pymongo
python-nmap
python-dotenv
pyshark
psutil
